注释设置自增id代码

+++ ocean-provider/ocean-mes/ocean-mes-sfc/src/main/java/com/fantechs/provider/mes/sfc/mapper/MesSfcWorkOrderBarcodeMapper.java
@@ -62,7 +62,7 @@ public interface MesSfcWorkOrderBarcodeMapper extends MyMapper<MesSfcWorkOrderBa
 
     List<SyncWorkOrderBarcodeDto> selectPartField(Map<String, Object> map);
 
-    Long selectMaxWorkOrderBarcodeId();
+    int selectMaxWorkOrderBarcodeId();
 
     void setAutoIncrement(Integer autoIncrementId);
 }
